tools/weixin-favor-kb/requirements.txt
faster-whisper>=1.1.0
opencv-python-headless>=4.9.0
rapidocr-onnxruntime>=1.3.0
openai>=1.30.0
pyyaml>=6.0
jinja2>=3.1.0
numpy>=1.26.0
loguru>=0.7.0
rich>=13.0.0
